app/requirements.txt
streamlit==1.44.1
pymupdf==1.25.5
python-docx==1.1.2
scikit-learn==1.6.1
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
selenium==4.29.0
webdriver-manager==4.0.2
plotly==5.24.1
pandas==2.2.3
